+++ demos/browser/app/meetingV2/meetingV2.ts
@@ -36,6 +36,7 @@ import {
   VideoTileState,
   ClientVideoStreamReceivingReport,
 } from '../../../../src/index';
+import WebRTCStatsCollector from './webrtcstatscollector/WebRTCStatsCollector';
 
 class DemoTileOrganizer {
   // this is index instead of length
@@ -149,6 +150,7 @@ export class DemoMeetingApp implements AudioVideoObserver, DeviceChangeObserver,
     'button-speaker': true,
     'button-content-share': false,
     'button-pause-content-share': false,
+    'button-video-stats': false,
   };
 
   contentShareType: ContentShareType = ContentShareType.ScreenCapture;
@@ -162,6 +164,9 @@ export class DemoMeetingApp implements AudioVideoObserver, DeviceChangeObserver,
   lastMessageSender: string | null = null;
   lastReceivedMessageTimestamp = 0;
   meetingEventPOSTLogger: MeetingSessionPOSTLogger;
+
+  hasChromiumWebRTC: boolean = this.defaultBrowserBehaviour.hasChromiumWebRTC();
+  statsCollector: WebRTCStatsCollector = new WebRTCStatsCollector();
 
   constructor() {
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -451,6 +456,14 @@ export class DemoMeetingApp implements AudioVideoObserver, DeviceChangeObserver,
       });
     });
 
+    const buttonVideoStats = document.getElementById('button-video-stats');
+    buttonVideoStats.addEventListener('click', () => {
+      if (this.isButtonOn('button-video-stats')) {
+        document.querySelectorAll('.stats-info').forEach(e => e.remove());
+      }
+      this.toggleButton('button-video-stats');
+    });
+
     const sendMessage = () => {
       new AsyncScheduler().start(() => {
         const textArea = document.getElementById('send-message') as HTMLTextAreaElement;
@@ -655,6 +668,48 @@ export class DemoMeetingApp implements AudioVideoObserver, DeviceChangeObserver,
       (document.getElementById('video-downlink-bandwidth') as HTMLSpanElement).innerText =
         'Available Downlink Bandwidth: Unknown';
     }
+
+    this.hasChromiumWebRTC &&
+    this.isButtonOn('button-video-stats') &&
+    this.getAndShowWebRTCStats();
+  }
+
+  getAndShowWebRTCStats() {
+    const videoTiles = this.audioVideo.getAllVideoTiles();
+    if (videoTiles.length === 0) {
+      return;
+    }
+    for (const videoTile of videoTiles) {
+      const tileState = videoTile.state();
+      if (tileState.paused || tileState.isContent) {
+        continue;
+      }
+      const tileId = videoTile.id();
+      const tileIndex = this.tileIdToTileIndex[tileId];
+      this.getStats(tileIndex);
+      if (tileState.localTile) {
+        this.statsCollector.showUpstreamStats(tileIndex);
+      } else {
+        this.statsCollector.showDownstreamStats(tileIndex);
+      }
+    }
+
+  }
+
+  async getStats(tileIndex: number) {
+    const id = `video-${tileIndex}`;
+    const videoElement = document.getElementById(id) as HTMLVideoElement;
+    if (!videoElement || !videoElement.srcObject) {
+      return;
+    }
+
+    const stream = videoElement.srcObject as MediaStream;
+    const tracks = stream.getVideoTracks();
+    if (tracks.length === 0) {
+      return;
+    }
+    const report = await this.audioVideo.getRTCPeerConnectionStats(tracks[0]);
+    this.statsCollector.processWebRTCStatReportForTileIndex(report, tileIndex);
   }
 
   async createLogStream(configuration: MeetingSessionConfiguration, pathname: string): Promise<void> {
@@ -674,7 +729,7 @@ export class DemoMeetingApp implements AudioVideoObserver, DeviceChangeObserver,
       this.log(error.message);
     }
   }
-  
+
   eventDidReceive(name: EventName, attributes: EventAttributes): void {
     this.log(`Received an event: ${JSON.stringify({ name, attributes })}`);
     const { meetingHistory, ...otherAttributes } = attributes;
@@ -776,6 +831,7 @@ export class DemoMeetingApp implements AudioVideoObserver, DeviceChangeObserver,
   }
 
   leave(): void {
+    this.statsCollector.resetStats();
     this.audioVideo.stop();
     this.roster = {};
   }
@@ -1451,6 +1507,7 @@ export class DemoMeetingApp implements AudioVideoObserver, DeviceChangeObserver,
     const tileElement = document.getElementById(`tile-${tileIndex}`) as HTMLDivElement;
     const videoElement = document.getElementById(`video-${tileIndex}`) as HTMLVideoElement;
     const nameplateElement = document.getElementById(`nameplate-${tileIndex}`) as HTMLDivElement;
+    const attendeeIdElement = document.getElementById(`attendeeid-${tileIndex}`) as HTMLDivElement;
     const pauseButtonElement = document.getElementById(`video-pause-${tileIndex}`) as HTMLButtonElement;
 
     pauseButtonElement.addEventListener('click', () => {
@@ -1468,7 +1525,7 @@ export class DemoMeetingApp implements AudioVideoObserver, DeviceChangeObserver,
     this.tileIndexToTileId[tileIndex] = tileState.tileId;
     this.tileIdToTileIndex[tileState.tileId] = tileIndex;
     this.updateProperty(nameplateElement, 'innerText', tileState.boundExternalUserId.split('#')[1]);
-
+    this.updateProperty(attendeeIdElement, 'innerText', tileState.boundAttendeeId);
     this.showTile(tileElement, tileState);
     this.updateGridClasses();
     this.layoutFeaturedTile();