+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Add github actions continuous integration workflow and deploy workflow
+- [Demo] Show video WebRTC stats and attendeeId on video tile hover
 
 ### Changed
 - [Documentation] Updated HTTP to HTTPS in README URL links